# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.AccountResponse:
    properties:
      created_at:
        type: string
      display_name:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      timezone:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.BankAccount:
    properties:
      account_id:
        description: ссылка на главный аккаунт
        type: integer
      account_type:
        description: '"cash", "debit", "credit", "savings"'
        type: string
      bank_name:
        description: '"Kaspi", "Halyk", "Cash"'
        type: string
      created_at:
        type: string
      currency:
        description: '"KZT", "USD", "EUR"'
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        description: '"Каспи рубли", "Халык доллары", "Наличные USD"'
        type: string
      updated_at:
        type: string
    type: object
...
  models.CreateAccountRequest:
    properties:
      display_name:
        maxLength: 40
        minLength: 2
        type: string
    required:
    - display_name
    type: object
  models.CreateBankAccountRequest:
    properties:
      account_type:
        enum:
        - cash
        - debit
        - credit
        - savings
        type: string
      bank_name:
        maxLength: 40
        minLength: 2
        type: string
      currency:
        enum:
        - KZT
        - USD
        - EUR
        - RUB
        type: string
      name:
        maxLength: 40
        minLength: 2
        type: string
    required:
    - account_type
    - bank_name
    - currency
    - name
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Personal Finance Management API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fin-Core API
  version: "1.0"
paths:
  /account:
    get:
      description: Get the authenticated user's account information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AccountResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Account not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user account
      tags:
      - accounts
    post:
      consumes:
      - application/json
      description: Create a new financial account for the authenticated user
      parameters:
      - description: Account creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AccountResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new account
      tags:
      - accounts
...
  /bankAccounts:
    get:
      description: Get all bank accounts for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BankAccount'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all bank accounts
      tags:
      - bank-accounts
    post:
      consumes:
      - application/json
      description: Create a new bank account for the authenticated user
      parameters:
      - description: Bank account creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateBankAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.BankAccount'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new bank account
      tags:
      - bank-accounts
  /bankAccounts/{bank_account_id}:
    delete:
      description: Delete a bank account by ID (will also delete all related transactions)
      parameters:
      - description: Bank Account ID
        in: path
        name: bank_account_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Access denied
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a bank account
      tags:
      - bank-accounts
    get:
      description: Get a specific bank account by ID for the authenticated user
      parameters:
      - description: Bank Account ID
        in: path
        name: bank_account_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BankAccount'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Bank account not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a specific bank account
      tags:
      - bank-accounts
  /bankAccounts/{bank_account_id}/activate:
    put:
      description: Activate a previously deactivated bank account by ID
      parameters:
      - description: Bank Account ID
        in: path
        name: bank_account_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Access denied
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Activate a bank account
      tags:
      - bank-accounts
  /bankAccounts/{bank_account_id}/deactivate:
    put:
      description: Deactivate a bank account by ID (keeps transactions but hides account)
      parameters:
      - description: Bank Account ID
        in: path
        name: bank_account_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Access denied
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Deactivate a bank account
      tags:
      - bank-accounts
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"